[WIPPETYWIP] start major refactoring (doesn't compile)
